==> 03_LatPrecip_GLMM_analysis/03_genotype_clusters.Rmd <==
---
title: "Calculating clusters"
author: "FiG-T"
date: "2024-11-11"
output:
 html_document:
   keep_md: true
   toc: true
toc: true
editor_options:
  markdown:
    wrap: 80
  chunk_output_type: inline
---

# Calculating clusters

```{r libraries}
library(dplyr)
```

As traditional LD is not easily calculable for non-recombining mtDNA, we are
using r2 instead. We use this based on the plain genotype data for the 209 SNPs
used in the analysis for the down-sampled number of individuals used.

```{r import_data}

gt_meta_ds_subset <- feather::read_feather(
  path = "data/feather/gt_meta_pc_ds_subset_08_2024.feather"  
)

nrow(gt_meta_ds_subset)
```

```{r calculate_SNP_cor}

# convert genotypes to matrix
GT_cor <- Hmisc::rcorr(
  as.matrix(
    gt_meta_ds_subset[,c(24:ncol(gt_meta_ds_subset))]
  )
)

GT_cor <- reshape2::melt(
  GT_cor$r,
  na.rm = TRUE
)

```

```{r change_labels}

# to match the rCRS positions
GT_cor_pos <- GT_cor %>%
  filter( 
    str_detect(Var1, "pos_.*")
  ) %>%
  select(
    Var1
  ) %>%
  tidyr::separate(
    Var1, 
    c("text","pos"), 
    sep = "_"
  ) %>%
  distinct

# convert values to numeric
GT_cor_pos$pos <- as.numeric(GT_cor_pos$pos)

# join with the phased positions
GT_cor_pos <- left_join(
  GT_cor_pos, 
  rCRS, 
  by = c("pos" = "aln_pos")
)

# add in the prefix again
GT_cor_pos$rCRS_pos <- stringr::str_c(
  "pos_", 
  GT_cor_pos$rCRS_pos
)
GT_cor_pos$pos <- stringr::str_c(
  "pos_", 
  GT_cor_pos$pos
)
 
# replace the names with the phased names 
for(i in seq_len(nrow(GT_cor_pos))) {
  GT_cor <- GT_cor %>%
    mutate(
      Var1 = str_replace_all(
        Var1, 
        pattern = GT_cor_pos$pos[i], 
        replacement = GT_cor_pos$rCRS_pos[i]
      )
    )
  
  GT_cor <- GT_cor %>%
    mutate(
      Var2 = str_replace_all(
        Var2, 
        pattern = GT_cor_pos$pos[i], 
        replacement = GT_cor_pos$rCRS_pos[i]
      )
    )
}

GT_cor_env <- GT_cor %>%
  filter( ! str_detect(Var1, "pos_.*")) %>%
  select(Var1) %>%
  distinct() 

GT_cor_env <- c(GT_cor_env$Var1)

GT_cor$Var1 <- factor(
  x = GT_cor$Var1, 
  levels = c(
    GT_cor_env,
    c(GT_cor_pos$rCRS_pos)
  )
)

GT_cor$Var2 <- factor(
  x = GT_cor$Var2, 
  levels = c(
    GT_cor_env,
    c(GT_cor_pos$rCRS_pos)
  )
)
```

```{r plot_correlations}
ggplot2::ggplot(
  data = GT_cor,
  ggplot2::aes(
    x = Var2,
    y = Var1,
    fill = value
  )
) +
  geom_tile(
    colour = "white"
  ) +
#  geom_text(
#    mapping = aes(
#      label = value
#    ), 
#    #size = 1.25 # if saving to pdf
#    size = 2.5
#  ) +
  xlab("")+
  ylab("")+
  scale_fill_gradient2(
    low = "deepskyblue4",
    high = "darkorchid3",
    mid = "snow3",
    midpoint = 0,
    #limit = c(0.5,1),
    #space = "Lab",
    #name="Pearson\nCorrelation"
  ) +
  labs(fill = "R-squared Correlation Coefficient") +
  theme(
    axis.text.x = element_text(
      angle = 90,
      vjust = 0.25,
      hjust = 0.5,
      #size = 12, 
      size = 6
    ), 
    axis.text.y = element_text(
      angle = 0,
      vjust = 0.5,
      hjust = 0.5, 
      #size = 12, 
      size = 6
    ), 
    legend.position = "bottom"
    # axis.title = element_blank(), 
    # legend.position = "none"
  )

```

```{r linked_hits}

linked_snps <-GT_cor %>%
  filter(value != 1) %>%
  filter(value >= 0.2 | value <= -0.2) %>%
  group_by(Var1) %>%
  reframe(N = n())

unique(linked_snps$Var1)

# recalulate matrix
GT_cor_clust <- Hmisc::rcorr(
  as.matrix(
    gt_meta_ds_subset[,c(24:ncol(gt_meta_ds_subset))]
  )
)
GT_cor_clust <- GT_cor_clust$r

GT_clust_dist <- 1 - GT_cor_clust

GT_clust_dist <- hclust(
  as.dist(
    GT_clust_dist
  ), 
  method = "complete"
)

GT_clusters <- cutree(
  GT_clust_dist, 
  h = 0.8 # values that are more than 0.2 correlated will be clustered together
)

length(unique(GT_clusters))

# add in phased labels again
for(i in c(1:nrow(GT_cor_pos))) {
   GT_clust_dist$labels <- str_replace_all(
        GT_clust_dist$labels, 
        pattern = GT_cor_pos$pos[i], 
        replacement = GT_cor_pos$rCRS_pos[i]
      )
}

plot(
  GT_clust_dist, 
  #xlab = ""
  )

GT_clust_names <- data.frame(
  pos = colnames(GT_cor_clust),
  cluster = GT_clusters
)

for(i in c(1:nrow(GT_cor_pos))) {
   GT_clust_names$pos <- str_replace_all(
        GT_clust_names$pos, 
        pattern = GT_cor_pos$pos[i], 
        replacement = GT_cor_pos$rCRS_pos[i]
      )
}

GT_clust_names$rCRS <- str_remove_all(
  GT_clust_names$pos, 
  "[:alpha:]|[:punct:]"
)
GT_clust_names$rCRS <- as.numeric(GT_clust_names$rCRS)

```

```{r save_clusters}

GT_clust_names <- GT_clust_names %>%
  select(
    rCRS, cluster
  )

feather::write_feather(
  x = GT_clust_names,
  path = "results/genotype_correlated_clusters_r0.2.feather"
)
```
